+++ src/main/java/net/etylop/immersivefarming/utils/integration/jei/JEIHelper.java
@@ -1,0 +1,107 @@
+/*
+ * BluSunrize
+ * Copyright (c) 2017
+ *
+ * This code is licensed under "Blu's License of Common Sense"
+ * Details can be found in the license file in the root folder of this project
+ */
+
+package net.etylop.immersivefarming.utils.integration.jei;
+
+
+import blusunrize.immersiveengineering.api.crafting.IJEIRecipe;
+import blusunrize.immersiveengineering.api.crafting.cache.CachedRecipeList;
+import blusunrize.immersiveengineering.common.util.compat.jei.IEFluidTooltipCallback;
+import mezz.jei.api.IModPlugin;
+import mezz.jei.api.JeiPlugin;
+import mezz.jei.api.gui.drawable.IDrawableStatic;
+import mezz.jei.api.gui.ingredient.IRecipeSlotTooltipCallback;
+import mezz.jei.api.helpers.IGuiHelper;
+import mezz.jei.api.registration.*;
+import mezz.jei.api.runtime.IJeiRuntime;
+import net.etylop.immersivefarming.ImmersiveFarming;
+import net.etylop.immersivefarming.api.crafting.ComposterRecipe;
+import net.etylop.immersivefarming.block.IFBlocks;
+import net.etylop.immersivefarming.gui.ComposterScreen;
+import net.minecraft.client.Minecraft;
+import net.minecraft.resources.ResourceLocation;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.crafting.Recipe;
+
+import java.util.List;
+import java.util.function.Predicate;
+
+@JeiPlugin
+public class JEIHelper implements IModPlugin {
+	private static final ResourceLocation UID = new ResourceLocation(ImmersiveFarming.MOD_ID, "main");
+	public static final ResourceLocation JEI_GUI = new ResourceLocation(ImmersiveFarming.MOD_ID, "textures/gui/jei_elements.png");
+	public static IDrawableStatic slotDrawable;
+	public static IRecipeSlotTooltipCallback fluidTooltipCallback = new IEFluidTooltipCallback();
+
+	@Override
+	public ResourceLocation getPluginUid() {
+		return UID;
+	}
+
+	@Override
+	public void registerItemSubtypes(ISubtypeRegistration subtypeRegistry) {
+	}
+
+	@Override
+	public void registerIngredients(IModIngredientRegistration registry) {
+	}
+
+	@Override
+	public void registerCategories(IRecipeCategoryRegistration registry) {
+		//Recipes
+		IGuiHelper guiHelper = registry.getJeiHelpers().getGuiHelper();
+		registry.addRecipeCategories(
+				new ComposterRecipeCategory(guiHelper)
+		);
+
+		slotDrawable = guiHelper.getSlotDrawable();
+	}
+
+	@Override
+	public void registerVanillaCategoryExtensions(IVanillaCategoryExtensionRegistration registration) {
+
+	}
+
+	@Override
+	public void registerRecipes(IRecipeRegistration registration) {
+		registration.addRecipes(ComposterRecipeCategory.TYPE, filter(getRecipes(ComposterRecipe.RECIPES), IJEIRecipe::listInJEI));
+	}
+
+	private <T> List<T> filter(List<T> in, Predicate<T> include) {
+		return in.stream().filter(include).toList();
+	}
+
+	private <T extends Recipe<?>> List<T> getRecipes(CachedRecipeList<T> cachedList) {
+		return List.copyOf(cachedList.getRecipes(Minecraft.getInstance().level));
+	}
+
+	@Override
+	public void registerRecipeTransferHandlers(IRecipeTransferRegistration registration) {
+	}
+
+	@Override
+	public void registerRecipeCatalysts(IRecipeCatalystRegistration registration) {
+		registration.addRecipeCatalyst(new ItemStack(IFBlocks.COMPOSTER.get()), ComposterRecipeCategory.TYPE);
+	}
+
+	@Override
+	public void registerGuiHandlers(IGuiHandlerRegistration registration) {
+		registration.addRecipeClickArea(ComposterScreen.class, 52, 11, 16, 47, ComposterRecipeCategory.TYPE);
+	}
+
+
+	@Override
+	public void registerAdvanced(IAdvancedRegistration registration) {
+
+	}
+
+	@Override
+	public void onRuntimeAvailable(IJeiRuntime jeiRuntime) {
+	}
+
+}