+++ changelog.md
@@ -2,6 +2,7 @@
 
 * Crops cannot grow in cold weather anymore (cold biomes or serene season winter)
 * Changed the sprinkler models
+* Composter multiblock
 
 ##### Version 1.18.2-0.1